// packages/webapp-libs/webapp-emails/src/types.ts
import { ComponentType } from 'react';

export enum EmailTemplateType {
  ACCOUNT_ACTIVATION = 'ACCOUNT_ACTIVATION',
  PASSWORD_RESET = 'PASSWORD_RESET',
  SUBSCRIPTION_ERROR = 'SUBSCRIPTION_ERROR',
  TRIAL_EXPIRES_SOON = 'TRIAL_EXPIRES_SOON',
  USER_EXPORT = 'USER_EXPORT',
  USER_EXPORT_ADMIN = 'USER_EXPORT_ADMIN',
  TENANT_INVITATION = 'TENANT_INVITATION',
  INVOICE_REQUEST_ASSIGNED = 'INVOICE_REQUEST_ASSIGNED',
  INVOICE_REQUEST_COMMENT = 'INVOICE_REQUEST_COMMENT',
  INVOICE_REQUEST_MENTION = 'INVOICE_REQUEST_MENTION',
  INVOICE_CREATED = 'INVOICE_CREATED',
  INVOICE_FILE_ADDED = 'INVOICE_FILE_ADDED',
  PROJECT_NOTE_MENTION = 'PROJECT_NOTE_MENTION',
  BACKUP_READY = 'BACKUP_READY',
  TENANT_DELETED = 'TENANT_DELETED',
  SOCIAL_ACCOUNT_LINKED = 'SOCIAL_ACCOUNT_LINKED',
  OTP_ENABLED = 'OTP_ENABLED',
  OTP_DISABLED = 'OTP_DISABLED',
  SIGNUP_GUIDANCE = 'SIGNUP_GUIDANCE',
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  OTP_REPLACED = 'OTP_REPLACED',
  PASSWORD_SET = 'PASSWORD_SET',
  //<-- INJECT EMAIL TYPE -->
}

export type EmailComponentProps = {
  to: string;
};

export type EmailTemplateDefinition = {
  Template: ComponentType<any>;
  Subject: ComponentType<any>;
};